# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для системы поиска и управления историческими хрониками",
        "title": "ChronicleSearch API",
        "contact": {
            "name": "API Support",
            "email": "support@chroniclesearch.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/chronicle_resources": {
            "get": {
                "description": "Получить список всех хроник с возможностью фильтрации по названию, автору и локации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Получить список хроник",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по автору",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по локации",
                        "name": "location",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание новой хроники (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Создать хронику",
                "parameters": [
                    {
                        "description": "Данные хроники",
                        "name": "chronicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Lab1_intermal_app_ds.ChronicleResource"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/chronicle_resources/{id_chronicle_resource}": {
            "get": {
                "description": "Получить детальную информацию о хронике по идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Получить хронику по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "id_chronicle_resource",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление существующей хроники (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Обновить хронику",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "id_chronicle_resource",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные хроники",
                        "name": "chronicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Lab1_intermal_app_ds.ChronicleResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление хроники по ID (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Удалить хронику",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "id_chronicle_resource",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/chronicle_resources/{id_chronicle_resource}/add_to_chronicle_request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавление хроники в черновик заявки (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Добавить хронику в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "id_chronicle_resource",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/chronicle_resources/{id_chronicle_resource}/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загрузка изображения для хроники (требуется авторизация)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_resources"
                ],
                "summary": "Загрузить изображение хроники",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "id_chronicle_resource",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Lab1_intermal_app_ds.ChronicleResource": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "date_of_creation": {
                    "type": "string"
                },
                "detailed_description": {
                    "type": "string"
                },
                "detailed_editions": {
                    "type": "string"
                },
                "detailed_significance": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "time_of_action": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {токен}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}